+++ script/repl.sh
@@ -2,4 +2,5 @@
 set -o errexit -o nounset -o pipefail
 cd "`dirname $0`/.."
 
-clj -A:test:bench -J--add-opens=java.base/java.io=ALL-UNNAMED -X clojure.core.server/start-server :name repl :port 5555 :accept clojure.core.server/repl :server-daemon false
+echo "Starting Socket REPL server on port 5555"
+clj -A:1.11.1:dev:test:bench -X clojure.core.server/start-server :name repl :port 5555 :accept clojure.core.server/repl :server-daemon false